# server/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Housing Rental API
  version: 1.0.0
  description: API для работы с предложениями, отзывами и пользователями
servers:
  - url: http://localhost:5000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /static/{path}:
    get:
      summary: Получить изображение
      tags:
        - Offers
      responses:
        "200":
          description: Запрос выполнен успешно
        "404":
          description: Такое изображение не найдено
      parameters:
        - name: path
          in: path
          required: true
          schema:
            type: string
          example: images/apartment-123.jpg
          description: Путь к изображению
...